Moved GitHub Login code into LoginScreen.js
